The final PCB design
This is the final PCB. It is a custom PCB made on a software called Fritzing. We make our PCBs manually so we print these PDFs on glossy paper, then we iron it (like ironing clothes xD) on copper sheets that have a sheet of plastic under it. The ironed sheet is then placed in acid for an amount of time then removed when all the copper has been dissolved except for the copper under the ink that came of the paper, which forms the tracks for the PCB.