Final PCB Design
This is the PCB design that we have managed to make as small as possible to occupy less space, thus making the robot smaller. The PCB is mainly an Arduino Nano shield and its purpose is to make the electrical system more stable, as using jumpers to connect the components is not as as stable as a PCB because of the use of data cables in a PCB that can not be removed easily from the data terminal that is soldered into the PCB. The 3 pin terminals are for the IR sensors and the 5 pin terminal is for the Cytron motor driver.